requirements.txt
# Flask==0.12.1
# Flask_Cors==3.0.3
# flask-cors==3.0.3
# requests==2.13.0
Flask
Flask_Cors
flask-cors
requests
mysql-connector-python
